requirements.txt
fastapi
tensorflow
pillow
numpy
uvicorn
python-multipart
jinja2
